requirements.txt
Django==2.2.3
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.2
Pillow==6.1.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
